Move around definitions
